Remove duplicate code by creating isValidPrinter function
