+++ src/win32/get-printers.js
@@ -1,6 +1,7 @@
 "use strict";
 
 const execFileAsync = require("../utils/exec-file-async");
+const isValidPrinter = require("../utils/windows-printer-valid");
 
 const getPrinters = () => {
   const stdoutHandler = (stdout) => {
@@ -11,24 +12,7 @@ const getPrinters = () => {
       .map((printer) => printer.trim())
       .filter((printer) => !!printer)
       .forEach((printer) => {
-        const printerData = {
-          deviceId: null,
-          name: null,
-        };
-
-        const isValid = printer.split(/\r?\n/).some((line) => {
-          const [label, value] = line.split(":").map((el) => el.trim());
-
-          const lowerLabel = label.toLowerCase();
-
-          if (lowerLabel === "deviceid") printerData.deviceId = value;
-
-          if (lowerLabel === "name") printerData.name = value;
-
-          if (printerData.deviceId && printerData.name) return true;
-
-          return false;
-        });
+        const { isValid, printerData } = isValidPrinter(printer);
 
         if (!isValid) return;
